requirements.txt
fastapi==0.78.0
SQLAlchemy==1.4.39
uvicorn==0.18.2
pymysql
python-dotenv
pylint
pytest
httpx
pytest-asyncio
coverage
python-jose
passlib
python-multipart
datetime
requests
psycopg2-binary
PyJWT
gunicorn==20.1.0





